+++ MANIFEST.in
@@ -80,6 +80,7 @@ include psutil/arch/openbsd/cpu.c
 include psutil/arch/openbsd/disk.c
 include psutil/arch/openbsd/init.h
 include psutil/arch/openbsd/mem.c
+include psutil/arch/openbsd/pids.c
 include psutil/arch/openbsd/proc.c
 include psutil/arch/openbsd/socks.c
 include psutil/arch/openbsd/users.c
